Fixed v8 & chakra outputting ints instead of doubles for numbers
Former-commit-id: d711ec3d03368c5f6171b11cff7101e7fda87f60

+++ src/ChakraBind.cpp
@@ -687,7 +687,7 @@ double JSValueToNumber(JSContextRef Context,JSValueRef Value,JSValueRef* Excepti
 	return Double;
 }
 
-JSValueRef JSValueMakeNumber(JSContextRef Context,int IntValue)
+JSValueRef JSValueMakeNumber(JSContextRef Context,double IntValue)
 {
 	JsValueRef Value = nullptr;
 	auto Error = JsIntToNumber( IntValue, &Value );
